chore(site): remove unused frontend code
