+++ site/src/components/Sidebar/Sidebar.stories.tsx
@@ -1,14 +1,6 @@
 import type { Meta, StoryObj } from "@storybook/react-vite";
-import {
-	CalendarCogIcon,
-	FingerprintIcon,
-	KeyIcon,
-	LockIcon,
-	UserIcon,
-} from "lucide-react";
 import { Outlet } from "react-router";
-import { Avatar } from "#/components/Avatar/Avatar";
-import { Sidebar, SidebarHeader, SidebarNavItem } from "./Sidebar";
+import { SettingsSidebarNavItem, Sidebar } from "./Sidebar";
 
 const meta: Meta<typeof Sidebar> = {
 	title: "components/Sidebar",
@@ -31,26 +23,19 @@ export const Default: Story = {
 	],
 	render: () => (
 		<Sidebar>
-			<SidebarHeader
-				avatar={<Avatar fallback="Jon" />}
-				title="Jon"
-				subtitle="jon@coder.com"
-			/>
-			<SidebarNavItem href="account" icon={UserIcon}>
-				Account
-			</SidebarNavItem>
-			<SidebarNavItem href="schedule" icon={CalendarCogIcon}>
-				Schedule
-			</SidebarNavItem>
-			<SidebarNavItem href="security" icon={LockIcon}>
-				Security
-			</SidebarNavItem>
-			<SidebarNavItem href="ssh-keys" icon={FingerprintIcon}>
-				SSH Keys
-			</SidebarNavItem>
-			<SidebarNavItem href="tokens" icon={KeyIcon}>
-				Tokens
-			</SidebarNavItem>
+			<div className="flex flex-col gap-1">
+				<SettingsSidebarNavItem href="account">Account</SettingsSidebarNavItem>
+				<SettingsSidebarNavItem href="schedule">
+					Schedule
+				</SettingsSidebarNavItem>
+				<SettingsSidebarNavItem href="security">
+					Security
+				</SettingsSidebarNavItem>
+				<SettingsSidebarNavItem href="ssh-keys">
+					SSH Keys
+				</SettingsSidebarNavItem>
+				<SettingsSidebarNavItem href="tokens">Tokens</SettingsSidebarNavItem>
+			</div>
 		</Sidebar>
 	),
 	parameters: {
