+++ site/src/pages/AgentsPage/components/ChatSharingPopover.tsx
@@ -1,4 +1,4 @@
-import { EllipsisVerticalIcon, Share2Icon, UserPlusIcon } from "lucide-react";
+import { EllipsisVerticalIcon, UserPlusIcon } from "lucide-react";
 import { type FC, type ReactNode, useState } from "react";
 import { useMutation, useQuery, useQueryClient } from "react-query";
 import { toast } from "sonner";
@@ -18,12 +18,7 @@ import {
 	DropdownMenuItem,
 	DropdownMenuTrigger,
 } from "#/components/DropdownMenu/DropdownMenu";
-import { TopbarButton } from "#/components/FullPageLayout/Topbar";
-import {
-	Popover,
-	PopoverContent,
-	PopoverTrigger,
-} from "#/components/Popover/Popover";
+import { PopoverContent } from "#/components/Popover/Popover";
 import { Spinner } from "#/components/Spinner/Spinner";
 import {
 	Table,
@@ -40,12 +35,9 @@ import {
 	type UserOrGroupAutocompleteValue,
 } from "#/modules/workspaces/WorkspaceSharingForm/UserOrGroupAutocomplete";
 
-type ChatShareButtonProps = {
+type ChatSharingPopoverContentProps = {
 	chatId: string;
 	organizationId: string;
-};
-
-type ChatSharingPopoverContentProps = ChatShareButtonProps & {
 	open: boolean;
 };
 
@@ -400,38 +392,5 @@ export const ChatSharingPopoverContent: FC<ChatSharingPopoverContentProps> = ({
 				) : null}
 			</div>
 		</PopoverContent>
-	);
-};
-
-export const ChatShareButton: FC<ChatShareButtonProps> = ({
-	chatId,
-	organizationId,
-}) => {
-	const [open, setOpen] = useState(false);
-	const [contentGeneration, setContentGeneration] = useState(0);
-
-	const handleOpenChange = (nextOpen: boolean) => {
-		if (nextOpen) {
-			setContentGeneration((generation) => generation + 1);
-		}
-
-		setOpen(nextOpen);
-	};
-
-	return (
-		<Popover open={open} onOpenChange={handleOpenChange}>
-			<PopoverTrigger asChild>
-				<TopbarButton data-testid="chat-share-button">
-					<Share2Icon />
-					Share
-				</TopbarButton>
-			</PopoverTrigger>
-			<ChatSharingPopoverContent
-				key={contentGeneration}
-				chatId={chatId}
-				organizationId={organizationId}
-				open={open}
-			/>
-		</Popover>
 	);
 };
